MYM-94 leave error fix

+++ backend/back/src/socket/chat.gateway.ts
@@ -216,17 +216,19 @@ export class EventsGateway
 	handleDisconnect(@ConnectedSocket() socket: Socket) {
 		this.logger.log(`${socket.id} socket disconnected`);
 		this.logger.log(`${socket.data.roomList}`);
-		userList[socket.data.user.uid].status = 'offline';
-		socket.broadcast.emit("user-update", {
-			userId: socket.data.user.uid,
-			userDisplayName: socket.data.user.nickname.split('#', 2)[0],
-			userProfileUrl: socket.data.user.profileUrl,
-			userStatus: userList[socket.data.user.uid].status,
-		});
-		socket.data.roomList.map((roomId: number) => {
-			this.deleteRoomLogic(socket, roomId);
-		});
-		delete userList[socket.data.user.uid];
+		if (socket.data?.user?.uid !== undefined) {
+			userList[socket.data.user.uid].status = 'offline';
+			socket.broadcast.emit("user-update", {
+				userId: socket.data.user.uid,
+				userDisplayName: socket.data.user.nickname.split('#', 2)[0],
+				userProfileUrl: socket.data.user.profileUrl,
+				userStatus: userList[socket.data.user.uid].status,
+			});
+			socket.data.roomList.map((roomId: number) => {
+				this.deleteRoomLogic(socket, roomId);
+			});
+			delete userList[socket.data.user.uid];
+		}
 	}
 
 	@SubscribeMessage("clear-data")
